Move access-key to env
