+++ addresses/urls.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+from django.urls import path
+from . import views 
+urlpatterns=[
+    path("", views.getAddress.as_view()),#[POST, PUT, DELETE] 사용자가 등록한 동네 조회, 삭제, POST] +)+)동네 재설정 추가, 동네 삭제[PUT]
+    path("get/ip", views.getIP.as_view()), #user 현 위치의 동네 조회[GET]
+    path("get/query", views.getQuery.as_view()), #검색어 기반 동네 조회 [GET]
+]
